Dishify2/app/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
fastapi==0.54.1
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
h11==0.9.0
idna==2.9
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.5.1
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
starlette==0.13.2
text2num==2.2.0
urllib3==1.25.9
uvicorn==0.11.5
websockets==8.1
